requirements.txt
stable_baselines3
gym[all]
git+https://github.com/ejmichaud/mazelab.git
tqdm
h5py
PyYAML
numpy
matplotlib
scipy
scikit-image
jupyter
-e .
